+++ src/i18n/languages/enGB.ts
@@ -144,6 +144,7 @@ const dictionary: RawTranslationPackage = {
     ISTEXT: 'ISTEXT',
     LEFT: 'LEFT',
     LEN: 'LEN',
+    LINEST: 'LINEST',
     LN: 'LN',
     LOG10: 'LOG10',
     LOG: 'LOG',